requirements.txt
pip==23.2.1
attrs==22.2.0
cryptography==43.0.1
pytest==8.0.1
h11==0.14.0
typing_extensions==4.9.0
trio==0.24.0
outcome==1.3.0.post0
sniffio==1.3.0
sortedcontainers==2.4.0
cffi==1.16.0
idna==3.6
mysql-connector-python==9.0.0
Appium-Python-Client==1.3.0
selenium==4.1.0
urllib3==1.26.16
pluggy==1.4.0
iniconfig==2.0.0
certifi==2024.2.2
six==1.16.0
pycparser==2.21
websocket-client==1.8.0
eyes-selenium==5.26.1
marshmallow==2.21.0
packaging==23.2
wsproto==1.2.0
requests==2.31.0
python-dotenv==1.0.1